Workaround to make the role work when a single home directory is shared across users
UNIX users may share a home directory. An example case is the "operator" user bundled with some Linux distributions (https://unix.stackexchange.com/questions/11638/what-is-the-point-of-the-operator-user), which shares its home directory with root.

This role replaces the authorized keys file for all UNIX users that have one in their home directory, regardless of whether the user is referenced in `ssh_manager_authorized_persons` or not. In the latter case, the authorized keys file is replaced with an empty file.

This poses a problem whenever two UNIX users share the same home directory, in which case it is unclear what the contents of the authorized keys file should be.

As a workaround:
- Do not attempt to replace the authorized keys file for UNIX users that have one in their home directory but don't own it, unless explicitly referenced in ssh_manager_authorized_persons
- Fail when attempting to replace the authorized keys file for UNIX users that don't own the authorized keys file in their home directory

In addition, replace `keys` with `key` in `ssh_manager_authorized_persons` (the latter is reserved by Ansible to access a dictionary's keys and causes bugs).